Export permutation function and Header wrapper fields in index package
